Now ISpectrum is a necessary interface, Spectrum is an abstract class that needs to be extended in a weird way, and MzSpectrum is a fully specified class that can be instantiated
