+++ Spectra/Spectra.csproj
@@ -48,13 +48,14 @@
   <ItemGroup>
     <Compile Include="Peak\Peak.cs" />
     <Compile Include="Spectrum\IHasMassSpectrum.cs" />
+    <Compile Include="Spectrum\MzSpectrum.cs" />
     <Compile Include="Util\ClassExtensions.cs" />
     <Compile Include="Range\DoubleRange.cs" />
     <Compile Include="Range\IRange.cs" />
     <Compile Include="Spectrum\ISpectrum.cs" />
     <Compile Include="Peak\MzPeak.cs" />
     <Compile Include="Range\MzRange.cs" />
-    <Compile Include="Spectrum\MzSpectrum.cs" />
+    <Compile Include="Spectrum\Spectrum.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Range\Range.cs" />
     <Compile Include="Util\Tolerance.cs" />
